Add files via upload
Are you a parent who is trying to make ends meet? Maybe you get paid hourly and have some flexibility in your schedule (e.g., nurses).
This program is built to help you as a parent decide what is the right mix between working and days in daycare for your child(ren). Running this program and answering its questions will give you an idea for which mix works best for you - given your wages and costs of daycare.